varnostne zascite

+++ cities.php
@@ -2,7 +2,11 @@
 include_once "header.php";
 ?>
 <h1>Kraji</h1>
-<a href="city_add.php">DODAJ KRAJ</a>
+<?php
+    if (isAdmin()) {
+        echo '<a href="city_add.php">Dodaj kraj</a>';
+    }
+?>
 <ul>
     <?php
     include_once "db.php";
@@ -11,8 +15,10 @@ include_once "header.php";
     $stmt->execute();
     while ($row = $stmt->fetch()) {
         echo '<li>'.$row['title'];
-        echo ' <a href="city_delete.php?id='.$row['id'].'" onclick="return confirm(\'Prepričani?\')">(x)</a>';
-        echo ' <a href="city_edit.php?id='.$row['id'].'">(u)</a>';
+        if (isAdmin()) {
+            echo ' <a href="city_delete.php?id='.$row['id'].'" onclick="return confirm(\'Prepričani?\')">(x)</a>';
+            echo ' <a href="city_edit.php?id='.$row['id'].'">(u)</a>';
+        }
         echo '</li>';
     }
     ?>
