+++ city_delete.php
@@ -1,4 +1,10 @@
 <?php
+include_once 'session.php';
+if (!isAdmin()) {
+    header("Location: index.php");
+    die();
+}
+
 include_once 'db.php';
 $id = $_GET['id'];
 $sql = "DELETE FROM cities WHERE id = ?";
